pyproject.toml
[project]
name = "xnat-selenium-tests"
version = "0.1.0"
description = "Selenium-based end-to-end regression tests for XNAT deployments"
authors = [{name = "Automation", email = "automation@example.com"}]
requires-python = ">=3.10"
dependencies = [
    "pytest>=7.4",
    "selenium>=4.13",
    "pyyaml>=6.0",
]

[build-system]
requires = ["setuptools>=65"]
build-backend = "setuptools.build_meta"
